backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.0.2
numpy==2.2.5
faiss-cpu==1.10.0
sentence-transformers==4.1.0
requests==2.32.3
pydantic==2.11.3